// src/api/generated/model/contestTeamMemberAnalytics.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */

export interface ContestTeamMemberAnalytics {
  id: string;
  contest_team_member_id: string;
  name: string;
  email: string;
  score?: number;
  is_flagged?: boolean;
  flagged_reason?: string | null;
  started_at?: string | null;
  ended_at?: string | null;
  is_participated?: boolean;
  is_leader?: boolean;
}